Update README.md
I'm currently an education professional, but after almost a decade in this field, I'm looking for a career shift.  At the moment taking my first steps into technology and in love with it. I'm open to collaborating on projects in Python, wich is the language I'm studying the most at the moment.